refactor: simplify imports to use configs package directly
- Remove AppConfig class (use app_config from make_config instead)
- Update all builders to import from configs instead of configs.pydantic
- Maintain consistent import style with proper alignment
- Leverage __all__ exports from configs package
