+++ src/configs/builds/environment.py
@@ -5,9 +5,9 @@ This module defines the configuration builder for ThermurEnv, which creates
 a Hydra-compatible config that instantiates the environment with validated
 parameters from the EnvironmentConfig Pydantic model.
 """
-from configs.pydantic import EnvironmentConfig
-from hydra_zen        import builds, zen
-from thermur          import ThermurEnv
+from configs    import EnvironmentConfig
+from hydra_zen  import builds, zen
+from thermur    import ThermurEnv
 
 
 build_environment = builds(
