+++ src/configs/builds/optimizer.py
@@ -4,10 +4,10 @@ Hydra-zen builder for the AdamW optimizer.
 This module defines the configuration builder for the optimizer used in
 training the GNN policy through imitation learning.
 """
-from configs.pydantic import TrainConfig
-from hydra_zen        import builds, zen
-from omegaconf        import SI
-from torch.optim      import AdamW
+from configs      import TrainConfig
+from hydra_zen    import builds, zen
+from omegaconf    import SI
+from torch.optim  import AdamW
 
 
 build_optimizer = builds(
